requirements/requirements_dml.txt
PySide6>=6.6.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
torch-directml>=0.2.0
torch>=2.1.0
torchvision>=0.16.0
onnxruntime-directml>=1.16.0
scipy>=1.11.0
scikit-image>=0.21.0
imageio>=2.31.0
psutil>=5.9.0
imageio-ffmpeg>=0.4.9
kornia>=0.7.0
insightface>=0.7.3
gfpgan>=1.3.8
basicsr>=1.4.2
moviepy>=1.0.3
realesrgan>=0.3.0
pyvirtualcam>=0.11.0

